// src/api/generated/model/privateDashboardLayoutPrivateDashboardTypes.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * Cosy API
 * Management API for Cosy (Cost Optimised Server Yard).
 * OpenAPI spec version: v1.0
 */

export type PrivateDashboardLayoutPrivateDashboardTypes = typeof PrivateDashboardLayoutPrivateDashboardTypes[keyof typeof PrivateDashboardLayoutPrivateDashboardTypes];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PrivateDashboardLayoutPrivateDashboardTypes = {
  METRIC: 'METRIC',
  LOGS: 'LOGS',
  FREETEXT: 'FREETEXT',
} as const;